Add distribution documentation to README
We add information and workflow on how to manually distribute the
DCSO Portal Python SDK through the Python Package Index (PyPI).

We also update the requirements to the latest of `setuptools` and
`pdoc3`. Since these are dependencies that only matter when creating
distributions, we renamed the file `requirements.txt` as
`requirements_dist.txt`.